chore(sample-apps): move audio-health debug indicators from tutorial to dogfood
Reverts the AudioHealthBadge + RecoveryActions widgets from the tutorial's
App.tsx. The tutorial keeps the iOS WKWebView companion infra (?call_id=
override, URL sync, window.call debug) and goes back to its minimal layout.

Adds the same widgets to react-dogfood's ActiveCallHeader, gated by the
existing ?debug=true flag (useIsDebugMode):

- AudioHealthIndicator renders an inline chip with per-direction mic/spk
  status dots and the raw reason when ?debug=true; default users keep the
  existing notification + no-audio icon.
- AudioRecoveryButtons (Resume audio, audioSession=play-and-record) sit
  next to it, also gated by ?debug=true. Useful for verifying audio
  recovery on iOS WKWebView from a real React-dispatched click.
- Adds SCSS for the new debug chip, status dots, and recovery buttons.
